requirements.txt
flask==2.0.0
pillow==9.1.0
qrcode==7.3.1
